# api/src/routes/organism.ts
// Package Imports
import { Router } from "express";
import { Op } from "sequelize";

import { authRole } from "../middleware/auth.middleware";
import OrganismGroup from "../models/OrganismGroup";
import Organism from "../models/Organism";
import HttpException from "../exceptions/HttpException";

// Type Imports
import type { Request, Response, NextFunction } from "express";

// Type Declarations
interface IOrganismBody {
  name: string;
  organismGroupId: number;
}

type RouteRequest = Request<
  Record<"organismGroupId" | "organismId", string>,
  Record<"organismGroupId", string>,
  IOrganismBody
>;

// Logic
const router = Router();

router.get(
  "/",
  async (req: RouteRequest, res: Response, next: NextFunction) => {
    try {
      const { name } = req.body;

      const result = await Organism.findAll({
        attributes: {
          exclude: ["organismGroupId"],
        },
        where: name
          ? {
              name: {
                [Op.iLike]: name,
              },
            }
          : {},
        include: [{ model: OrganismGroup, as: "organismGroup" }],
      });

      return res.status(200).send(result);
    } catch (error) {
      next(error);
    }
  }
);

/*
 * Deprecated route due to not meeting the requirements of the program any longer
router.get(
  "/organismGroup",
  async (req: RouteRequest, res: Response, next: NextFunction) => {
    try {
      const { organismGroupId } = req.query;

      if (!organismGroupId || organismGroupId === "-1")
        throw new HttpException(
          400,
          "A valid Organism Group ID must be provided"
        );

      const result = await Organism.findAll({
        attributes: {
          exclude: ["organismGroupId"],
        },
        where: { organismGroupId: { [Op.eq]: organismGroupId as string } },
        include: [{ model: OrganismGroup, as: "organismGroup" }],
      });

      return res.status(200).send(result);
    } catch (error) {
      next(error);
    }
  }
);
*/

router.get(
  "/:organismId",
  async (req: RouteRequest, res: Response, next: NextFunction) => {
    const { organismId } = req.params;

    try {
      const result = await Organism.findByPk(organismId, {
        attributes: {
          exclude: ["organismGroupId"],
        },
        include: [{ model: OrganismGroup, as: "organismGroup" }],
      });

      return res.status(200).send(result);
    } catch (error) {
      next(error);
    }
  }
);

// From this point, only users with the "admin" role can use the following routes.
router.use(authRole("admin"));

router.post(
  "/",
  async (req: RouteRequest, res: Response, next: NextFunction) => {
    try {
      const { name } = req.body;

      const defaultOrganismGroup = (await OrganismGroup.findAll())[0];
      if (!defaultOrganismGroup) {
        throw new HttpException(
          404,
          "The requested Organism Group doesn't exist"
        );
      }

      if (!name) throw new HttpException(400, "Faltan parametros");

      const result = await Organism.create({ name });
      await defaultOrganismGroup.addOrganism(result);

      return res.status(201).send(
        await Organism.findByPk(result.id, {
          attributes: {
            exclude: ["organismGroupId"],
          },
          include: [{ model: OrganismGroup, as: "organismGroup" }],
        })
      );
    } catch (error) {
      console.error(error);
      next(error);
    }
  }
);

router.put(
  "/:organismId",
  async (req: RouteRequest, res: Response, next: NextFunction) => {
    try {
      const { organismId } = req.params;
      const { name } = req.body;

      if (!organismId) {
        throw new HttpException(400, "The Organism ID is missing as the param");
      }
      const result = await Organism.findByPk(organismId);

      if (!result) {
        throw new HttpException(404, "The requested Organism doesn't exist");
      }

      if (name) await result.update({ name });

      return res.status(200).send(
        await Organism.findByPk(result.id, {
          attributes: {
            exclude: ["organismGroupId"],
          },
          include: [{ model: OrganismGroup, as: "organismGroup" }],
        })
      );
    } catch (error) {
      next(error);
    }
  }
);

router.delete(
  "/:organismId",
  async (req: RouteRequest, res: Response, next: NextFunction) => {
    try {
      const { organismId } = req.params;

      if (!organismId) {
        throw new HttpException(400, "The Organism ID is missing as the param");
      }
      const result = await Organism.findByPk(organismId);

      if (!result) {
        throw new HttpException(404, "The requested Organism doesn't exist");
      }

      await result.destroy();

      res.status(200).send("The choosed Organism was disabled successfully");
    } catch (error) {
      next(error);
    }
  }
);

export default router;